windows-misc-fix-20061005
still tracking down issues when large numbers of temporary files
are created and deleted in the same directory from multiple
machines.

VNOVNODE must return CM_ERROR_BADFD

ensure that all calls to cm_GetCallback are made from cm_SyncOp.

add CM_SCACHESYNC_FORCECB logic to cm_SyncOp so that it can be
removed from cm_GetAccess

don't mix CM_SCACHESYNC_NEEDCALLBACK with CM_SCACHESYNC_STOREDATA
or CM_SCACHESYNC_FETCHDATA

+++ src/WINNT/afsd/cm_conn.c
@@ -75,23 +75,23 @@ void cm_InitConn(void)
 	    code = RegQueryValueEx(parmKey, "ConnDeadTimeout", NULL, NULL,
 				    (BYTE *) &dwValue, &dummyLen);
 	    if (code == ERROR_SUCCESS)
-                ConnDeadtimeout = dwValue;
+                ConnDeadtimeout = (unsigned short)dwValue;
 
 	    dummyLen = sizeof(DWORD);
 	    code = RegQueryValueEx(parmKey, "HardDeadTimeout", NULL, NULL,
 				    (BYTE *) &dwValue, &dummyLen);
 	    if (code == ERROR_SUCCESS)
-                HardDeadtimeout = dwValue;
+                HardDeadtimeout = (unsigned short)dwValue;
 	    afsi_log("HardDeadTimeout is %d", HardDeadtimeout);
 	    RegCloseKey(parmKey);
 	}
 
 	afsi_log("lanmanworkstation : SessTimeout %d", RDRtimeout);
 	if (ConnDeadtimeout == 0)
-	    ConnDeadtimeout = RDRtimeout / 2;
+	    ConnDeadtimeout = (unsigned short) (RDRtimeout / 2);
 	afsi_log("ConnDeadTimeout is %d", ConnDeadtimeout);
 	if (HardDeadtimeout == 0)
-	    HardDeadtimeout = RDRtimeout;
+	    HardDeadtimeout = (unsigned short) RDRtimeout;
 	afsi_log("HardDeadTimeout is %d", HardDeadtimeout);
 
 	osi_EndOnce(&once);
